politeiad: Move timestamp verify functions.
This diff moves the timestamp verification code from the `tstore`
package to the `backendv2` package to reduce the number of indirect
dependencies that clients must import.
